+++ RELEASE-NOTES.md
@@ -16,6 +16,7 @@
 * #85 Series.map flavor that returns a DataFrame
 * #86 Expose JdbcConnector DataSource for ease of integration
 * #87 DataFrame.addColumns with RowMapper
+* #88 Index.toSeries() 
 
 ## Release 0.6